Conflicts:
	scripts/cityprop.js
	scripts/spot.js
	scripts/spots_all.js
	scripts/spots_user.js
	templates/account-edit.phtml
	templates/cityprop.phtml
	templates/cityprops.phtml
	templates/spot.phtml
	templates/spots_user.phtml

+++ scripts/spot.js
@@ -19,6 +19,7 @@ var marker = null;
 
 var voted = false;
 var _user = 0;
+var _logged_in = false;
 
 var edit = false;
 
@@ -37,22 +38,32 @@ $(function(){
         }).addTo(map);
 });
 
-
-
 function getData(){
     $.ajax({
         type: "GET",
         dataType: "json",
         contentType: "application/json",
         cache: false,
-        data: null,
         url: _root + "/api/spots/" + spot_id,
         success: function(data){
-            voted = data.voted;
-            showSpotLocation(data[0].coords.split(" ")[0].substr(1), data[0].coords.split(" ")[1].substr(0, data[0].coords.split(" ")[1].length-2));
-            showData(data[0]);
+            console.log(data);
+            if (data.hasOwnProperty('logged_in') && !data.logged_in) 
+            {
+                _logged_in = false;
+                showSpotLocation(data[0].coords.split(" ")[0].substr(1), data[0].coords.split(" ")[1].substr(0, data[0].coords.split(" ")[1].length-2));
+                showData(data[0]);
+            }
+            else
+            {
+                _logged_in = true;
+                voted = data.voted;
+                showSpotLocation(data[0].coords.split(" ")[0].substr(1), data[0].coords.split(" ")[1].substr(0, data[0].coords.split(" ")[1].length-2));
+                showData(data[0]);
+
+            }
         },
         error: function(){
+            console.log("Spot not loaded");
             //window.location= _root + "/404";
         }
     });
@@ -76,20 +87,20 @@ function showSpotLocation(lat, lng){
 }
 
 function showData(data){
-	$("#spotTitle").html(data.description);
+	$("#spotTitle").html(htmlEncode(data.description));
     var s = "";
     s+= "<section>";
-    s+= "<h3 class='spot_bBtm''>Your spotting</h3><p>" + data.description +"</p>";
-    s+= "<h3 class='spot_bBtm'>Your suggestion</h3><p>" + data.proposed +"</p>";
-    s+= "<h3 class='spot_bBtm'>Created</h3><p>" + data.createddate +"</p>";
-    s+= "<p>Likes: <span id='up'>"+ data.upvotes +"</span></p>";
-	s+= "<p>Dislikes: <span id='down'>"+ data.downvotes +"</span></p>";
-    if (!voted) {
-        s+= "<input type='button' class='btnFL' id='upbtn' value='Fancy that!' onclick='voteup()' />";
+    s+= "<h3>Problem</h3><p>" + htmlEncode(data.description) +"</p>";
+    s+= "<h3>Proposal</h3><p>" + htmlEncode(data.proposed) +"</p>";
+    s+= "<h3>Created</h3><p>" + data.createddate +"</p>";
+    s+= "<p>Upvotes: <span id='up'>"+ data.upvotes +"</span></p>";
+    if (_logged_in && !voted) {
+        s+= "<input type='button' id='upbtn' value='Fancy that!' onclick='voteup()' />";
     }
     
-    if (!voted) {
-        s+= "<input type='button' class='btnFL' id='downbtn' value='I don&apos;t get it' onclick='votedown()' />";
+    s+= "<p>Downvotes: <span id='down'>"+ data.downvotes +"</span></p>";
+    if (_logged_in && !voted) {
+        s+= "<input type='button' id='downbtn' value='I do not get it..' onclick='votedown()' />";
     }
     
     s+= "<div id='user'>";
@@ -99,8 +110,10 @@ function showData(data){
     s+= "<div id='comments'>";
     s+= "<h2 class='spot_bBtm'>Comments</h2>";
     s+= "<div id='comm_inh'>Comments</div>";
-    s+= "<textarea id='comm_text' class='commTA' placeholder='Fill in your comment.'></textarea>";
-    s+= "<input type='button' class='btnAction' value='Comment' onclick='comment()' id='comment_enter' />";
+    if (_logged_in) {
+        s+= "<textarea id='comm_text' placeholder='Fill in your comment.'></textarea>";
+        s+= "<input type='button' value='Comment' onclick='comment()' id='comment_enter' />";
+    };
     s+= "</div>";
     s+= "</section>";
     $("#spot").html(s);
@@ -134,9 +147,9 @@ function parseUser(user)
 {
     var s = "";
     s+= "<section>";
-    s+= "<p>Your first name: "+ user.firstname +"</p>";
-    s+= "<p>Your surname: "+ user.surname +"</p>";
-    s+= "<p>Your avatar: </p>";
+    s+= "<p>firstname: "+ htmlEncode(user.firstname) +"</p>";
+    s+= "<p>surname: "+ htmlEncode(user.surname) +"</p>";
+    s+= "<p>avatar: </p>";
     s+= "<img src='"+ _root +"/uploads/"+ user.avatar +"' />";
     s+= "</section>";
     $("#user_inh").html(s);
@@ -149,12 +162,12 @@ function getComments(spot)
         dataType: "json",
         contentType: "application/json",
         cache: false,
-        data: null,
         url: _root + "/api/spots/" + spot + "/comments",
         success: function(data){
+           
             var comments = data.comments;
             _user = data.user;
-            parseComments(comments);
+            parseComments(comments); 
         },
         error: function(){
             $("#comm_inh").html("No comments for this spotting.");
@@ -166,9 +179,9 @@ function parseComments(comments)
 {
     s= "";
     for (i = 0; i < comments.length; i++) {
-        s+="<p id='"+ comments[i].comment_id +"-comment'><span class='text'>" + comments[i].text + " ";
-        s+="</span><span>By " + comments[i].firstname + " " + comments[i].surname + " <strong>" + comments[i].modifieddate +"</strong></span></p>";
-        if (_user = comments[i].user_id) {
+        s+="<p id='"+ comments[i].comment_id +"-comment'><span class='text'>" + htmlEncode(comments[i].text) + " ";
+        s+="</span><span>By " + htmlEncode(comments[i].firstname) + " " + htmlEncode(comments[i].surname) + " <strong>" + (comments[i].modifieddate == null?comments[i].createddate:comments[i].modifieddate) +"</strong></span></p>";
+        if (_logged_in && _user == comments[i].user_id) {
             s+= "<input type='button' value='edit comment' onclick='editComment("+ comments[i].comment_id +")' />";
             s+= "<input type='button' value='delete comment' onclick='deleteComment("+ comments[i].comment_id +")' />";
         }
@@ -189,17 +202,24 @@ function voteup()
         contentType: "application/json",
         cache: false,
         data: null,
-        url: _root + "/api/spots/" + spot_id + "/voteup/",
+        url: _root + "/api/spots/" + spot_id + "/voteup",
         success: function(data){
-            if (data != "voted") {
-                console.log("ok");
-                var n = $("#up").html();
-                var nn = parseInt(n);
-                nn++;
-                $("#up").html(nn);
+            if (data.hasOwnProperty('logged_in') && !data.logged_in) 
+            {
+                window.location = _root + "/login";
             }
-            $("#upbtn").remove();
-            $("#downbtn").remove();
+            else
+            {
+                if (data != "voted") {
+                    console.log("ok");
+                    var n = $("#up").html();
+                    var nn = parseInt(n);
+                    nn++;
+                    $("#up").html(nn);
+                }
+                $("#upbtn").remove();
+                $("#downbtn").remove();
+            }
             
         },
         error: function(){
@@ -215,18 +235,24 @@ function votedown()
         dataType: "json",
         contentType: "application/json",
         cache: false,
-        data: null,
-        url: _root + "/api/spots/" + spot_id + "/votedown/",
+        url: _root + "/api/spots/" + spot_id + "/votedown",
         success: function(data){
-            if (data != "voted") {
-                console.log("ok");
-                var n = $("#down").html();
-                var nn = parseInt(n);
-                nn++;
-                $("#down").html(nn);
+            if (data.hasOwnProperty('logged_in') && !data.logged_in) 
+            {
+                window.location = _root + "/login";
             }
-            $("#downbtn").remove();
-            $("#upbtn").remove();
+            else
+            {
+                if (data != "voted") {
+                    console.log("ok");
+                    var n = $("#down").html();
+                    var nn = parseInt(n);
+                    nn++;
+                    $("#down").html(nn);
+                }
+                $("#downbtn").remove();
+                $("#upbtn").remove();
+            }
         },
         error: function(){
             //window.location= _root + "/404";
@@ -263,11 +289,18 @@ function comment()
             data: dataS,
             url: url,
             success: function(data){
-                $("#comm_text").val('');
-                edit = false;
-                $("#comm_text").data('id', "");
-                $("#comment_enter").val('Comment');
-                getComments(spot_id);
+                if (data.hasOwnProperty('logged_in') && !data.logged_in) 
+                {
+                 window.location = _root + "/login";
+                }
+                else
+                {
+                    $("#comm_text").val('');
+                    edit = false;
+                    $("#comm_text").data('id', "");
+                    $("#comment_enter").val('Comment');
+                    getComments(spot_id);
+                }
             },
             error: function(xhr){
                 edit = false;
@@ -299,7 +332,14 @@ function deleteComment(id){
             data: JSON.stringify({"lol":1}),
             url: _root + "/api/spots/"+spot_id+"/comments/" + id,
             success: function(data){
-                getComments(spot_id);
+                if (data.hasOwnProperty('logged_in') && !data.logged_in) 
+                {
+                    window.location = _root + "/login";
+                }
+                else
+                {
+                    getComments(spot_id);
+                }
             },
             error: function(xhr){
                 console.log(xhr);
